# static/meta/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for LibRate, a social media cataloguing and reviewing service",
        "title": "LibRate",
        "contact": {
            "name": "MJH",
            "email": "TODO@flagship.instance"
        },
        "license": {
            "name": "GNU Affero General Public License v3",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "dev"
    },
    "basePath": "/api",
    "paths": {
...
        "/members/follow": {
            "post": {
                "description": "Send a follow request to another user. If the user has automatic follow request acceptance\nenabled, the follow request will be accepted immediately.",
                "consumes": [
                    "application/json"
                ],
                "summary": "Send a follow request to a member (user)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The requester's JWT. Contains encrypted claims to the webfinger",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The webfinger of the member to follow",
                        "name": "followee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Receive notifications for all content created by the followee",
                        "name": "notify",
                        "in": "body",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "description": "Show this account's reblogs in home timeline",
                        "name": "reblogs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "204": {
                        "description": "When the followee is already followed",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "403": {
                        "description": "When at least one party blocks the other",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            },
            "delete": {
                "description": "Unfollow a member or remove follower",
                "consumes": [
                    "application/json"
                ],
                "summary": "Unfollow a member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The follower's JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The webfinger of the member to unfollow",
                        "name": "followee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/members/follow/requests": {
            "get": {
                "description": "Get own received follow requests or sent follow requests",
                "consumes": [
                    "application/json"
                ],
                "summary": "Get follow requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The JWT of the member. Contains encrypted claims to webfinger",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Get own sent follow requests. If false, get received follow requests",
                        "name": "own",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseHTTP"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ResponseHTTP": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}